move flow-specific concerns to its own library
