+++ flow-plugin/build.gradle
@@ -1,18 +1,10 @@
-repositories {
-    mavenCentral()
-    maven {
-        url "https://oss.sonatype.org/content/repositories/snapshots"
-    }
-}
-
-apply plugin: 'com.android.application'
+apply plugin: 'com.android.library'
 
 android {
     compileSdkVersion 23
-    buildToolsVersion "22.0.1"
+    buildToolsVersion "23.0.1"
 
     defaultConfig {
-        applicationId "com.davidstemmer.screenplay.sample.simple"
         minSdkVersion 14
         targetSdkVersion 23
         versionCode 1
@@ -28,8 +20,8 @@ android {
 
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile project(':library')
-    compile project(':flow-plugin')
-    compile 'com.android.support:support-v4:23.0.1'
+    testCompile 'junit:junit:4.12'
     compile 'com.android.support:appcompat-v7:23.0.1'
+    compile project(':library')
+    compile 'com.squareup.flow:flow:0.12'
 }
