+++ flow-plugin/src/androidTest/java/davidstemmer/com/flowplugin/ApplicationTest.java
@@ -1,0 +1,13 @@
+package davidstemmer.com.flowplugin;
+
+import android.app.Application;
+import android.test.ApplicationTestCase;
+
+/**
+ * <a href="http://d.android.com/tools/testing/testing_android.html">Testing Fundamentals</a>
+ */
+public class ApplicationTest extends ApplicationTestCase<Application> {
+    public ApplicationTest() {
+        super(Application.class);
+    }
+}
