+++ flow-plugin/src/main/java/com/davidstemmer/flow/plugin/screenplay/ScreenplayDispatcher.java
@@ -1,4 +1,4 @@
-package com.davidstemmer.screenplay.flow;
+package com.davidstemmer.flow.plugin.screenplay;
 
 import android.app.Activity;
 import android.view.ViewGroup;
@@ -11,6 +11,7 @@ import java.util.Deque;
 
 import flow.Flow;
 import flow.History;
+
 
 /**
  * Created by weefbellington on 10/11/15.
